+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 1.150.0 (2025-08-02)
+
+Full Changelog: [v1.149.0...v1.150.0](https://github.com/StructifyAI/structify-node/compare/v1.149.0...v1.150.0)
+
+### Features
+
+* **api:** manual updates ([0ac5fcb](https://github.com/StructifyAI/structify-node/commit/0ac5fcbfacf1fd85cbdb2175144ede520db322ee))
+* **api:** manual updates ([2b8f6e2](https://github.com/StructifyAI/structify-node/commit/2b8f6e20ac3d82e874d09c269253cdbe25a90990))
+* Update from Structify backend changes ([bf021d1](https://github.com/StructifyAI/structify-node/commit/bf021d11155b83ec98a3d22afb7d5d7f9d352542))
+
+
+### Bug Fixes
+
+* add dataset name as a sdk param ([b3f2fc4](https://github.com/StructifyAI/structify-node/commit/b3f2fc4f29a0dcebf0f144afa72a6f8219fb6b2c))
+
 ## 1.149.0 (2025-07-31)
 
 Full Changelog: [v1.148.1...v1.149.0](https://github.com/StructifyAI/structify-node/compare/v1.148.1...v1.149.0)